requirements.txt
aniso8601==9.0.1
argon2-cffi==21.1.0
black==21.9b0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
pycparser==2.20
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
